# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Core API de Wallet",
        "description": "Esta es la documentación del core de la API de Wallet",
        "contact": {
            "email": "soporte@tuapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Permite registrar un nuevo usuario",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID del usuario",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Nombre del usuario",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "Correo del usuario",
                                                    "type": "string",
                                                    "format": "email"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error en el registro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El correo ya está en uso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Autenticar un usuario",
                "description": "Permite autenticar a un usuario y obtener un token de acceso",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "description": "Tipo de token",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Tiempo de expiración del token",
                                            "type": "integer"
                                        },
                                        "access_token": {
                                            "description": "Token de acceso",
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "description": "Token de refresco",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Login failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get paginated list of user transactions",
                "description": "Retrieve paginated list of transactions for the authenticated user",
                "operationId": "getUserTransactions",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of transactions per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Create a new transaction",
                "description": "Store a new transaction in the user's wallet",
                "operationId": "storeTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wallet/{id}": {
            "put": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Update a transaction",
                "description": "Update a transaction in the user's wallet",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Delete a transaction",
                "description": "Delete a transaction from the user's wallet",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/balances": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get user balance",
                "description": "Retrieve the current balance for the authenticated user",
                "operationId": "getBalance",
                "responses": {
                    "200": {
                        "description": "Current balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BalanceResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/download-excel": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Download user transactions as Excel",
                "description": "Download all user transactions as an Excel file",
                "operationId": "downloadExcel",
                "responses": {
                    "200": {
                        "description": "Excel file download"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginUserRequest": {
                "title": "LoginUserRequest",
                "description": "Login User Request",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "RegisterUserRequest": {
                "title": "RegisterUserRequest",
                "description": "Register User Request",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "StoreTransactionRequest": {
                "required": [
                    "amount",
                    "description",
                    "type"
                ],
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateTransactionRequest": {
                "required": [
                    "amount",
                    "description",
                    "type"
                ],
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "BalanceResource": {
                "properties": {
                    "userId": {
                        "type": "integer"
                    },
                    "totalIncome": {
                        "type": "number"
                    },
                    "incomeThisMonth": {
                        "type": "number"
                    },
                    "totalExpense": {
                        "type": "number"
                    },
                    "expenseThisMonth": {
                        "type": "number"
                    },
                    "generalBalance": {
                        "type": "number"
                    },
                    "balanceThisMonth": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "WalletCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WalletResource"
                        }
                    },
                    "pagination": {
                        "properties": {
                            "total": {
                                "type": "integer"
                            },
                            "count": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "current_page": {
                                "type": "integer"
                            },
                            "total_pages": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "WalletResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token Bearer JWT para acceder a este recurso",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Wallet",
            "description": "Wallet"
        }
    ]
}